sigstore, test: propagate rename
