+++ sigstore/_internal/sct.py
@@ -35,7 +35,7 @@ from cryptography.x509.certificate_transparency import (
 )
 from cryptography.x509.oid import ExtendedKeyUsageOID
 
-from sigstore._internal.trustroot import CTKeyring
+from sigstore._internal.trust import CTKeyring
 from sigstore._utils import (
     DERCert,
     KeyID,
